Move VimscriptParser to vim-engine
